ZNS: Let the device keep track of ZNS zones for us
Zones after "fla_open"
FlexAlloc only opens zones implicitly. This implies that the device will
transition zones automatically to closed if needed. closed zones still
count to the active resources and if more zones are open than the
defined (by the device) active resources we receive an error from xnvme
that we forward to the caller.

Zones on "fla_close"
We do not change state of zones when we close. We leave them to be
managed by the drive.

Zones on "fla_object_destroy"
This gives back the object to the pool as a usuable object and therefore
transitions the Zone to "empty" by sending a reset command

Zones on "fla_object_close"
In order to give the caller the power to "manage" the open/active
resources we add a new call that will transition all related zones of an
object to FULL.

Signed-off-by: Joel Granados <j.granados@samsung.com>