Test: Activate ZNS HW testing
When FLA_TEST_{MD,}_DEV is passed to the testing infra we have test
failing. This fixes those tests and allows FA to be tested on ZNS HW.

Rounding slab sizes to zone sizes.
Some tests just used an arbitrary slab size. This would not be possible
with ZNS as the slab could not end/start in the middle of a zone. We now
round up to the nearest zone size in the test infra.

When there is a write in the middle of an object which is housed in a
zone then we should expect a failure instead of expecting the
read/write to occur flawlessly.

Some tests like the slab were removed from running when in ZNS mode.
as they were incompatible due to the fact that the slab size would be
automatically rounded up.

Signed-off-by: Joel Granados <j.granados@samsung.com>